♻️ refactor(playground): extract examples to separate file
- Move EXAMPLE_CATEGORIES constant to examples.ts
- Remove duplicate type definition, import from examples.ts
- Update CONTRIBUTING.md to reflect current Rust version (1.93)
- Remove duplicate commit message format section from CONTRIBUTING.md
- Fix trailing comma formatting in Playground.tsx
