+++ packages/mq-playground/src/Playground.tsx
@@ -16,22 +16,12 @@ import {
   VscCheck,
   VscLoading,
 } from "react-icons/vsc";
+import { EXAMPLE_CATEGORIES, EXAMPLES } from "./examples";
 
 type SharedData = {
   code: string;
   markdown: string;
   options: mq.Options;
-};
-
-type ExampleCategory = {
-  name: string;
-  examples: {
-    name: string;
-    code: string;
-    markdown: string;
-    isUpdate: boolean;
-    format: mq.Options["inputFormat"];
-  }[];
 };
 
 const CODE_KEY = "mq-playground.code";
@@ -44,336 +34,15 @@ const SIDEBAR_VISIBLE_KEY = "mq-playground.sidebar-visible";
 const TABS_KEY = "mq-playground.tabs";
 const ACTIVE_TAB_ID_KEY = "mq-playground.active_tab_id";
 
-const EXAMPLE_CATEGORIES: ExampleCategory[] = [
-  {
-    name: "Basic Element Selection",
-    examples: [
-      {
-        name: "Hello World",
-        code: `# Hello world.
-select(.h || .list || .code) + " world"`,
-        markdown: `# Hello
-
-- Hello
-
-\`\`\`
-Hello
-\`\`\`
-`,
-        isUpdate: false,
-        format: "markdown",
-      },
-      {
-        name: "Extract heading",
-        code: `.h`,
-        markdown: `# Heading 1
-
-## Heading 2
-
-### Heading 3
-
-Some text here.
-`,
-        isUpdate: false,
-        format: "markdown",
-      },
-      {
-        name: "Extract table",
-        code: `.[1][]`,
-        markdown: `# Product List
-
-| Product | Category | Price | Stock |
-|---------|----------|-------|-------|
-| Laptop  | Electronics | $1200 | 45 |
-| Monitor | Electronics | $350 | 28 |
-| Chair   | Furniture | $150 | 73 |
-| Desk    | Furniture | $200 | 14 |
-| Keyboard | Accessories | $80 | 35 |
-
-| Product | Category | Price | Stock |
-|---------|----------|-------|-------|
-| Mouse   | Accessories | $25 | 50 |
-| Headphones | Electronics | $120 | 32 |
-| Bookshelf | Furniture | $180 | 17 |
-| USB Cable | Accessories | $12 | 89 |
-| Coffee Maker | Appliances | $85 | 24 |
-`,
-        isUpdate: false,
-        format: "markdown",
-      },
-      {
-        name: "Extract list",
-        code: `.[] | select(.list.level == 1)`,
-        markdown: `# Product List
-
-- Electronics
-  - Laptop: $1200
-  - Monitor: $350
-  - Headphones: $120
-- Furniture
-  - Chair: $150
-  - Desk: $200
-  - Bookshelf: $180
-- Accessories
-  - Keyboard: $80
-  - Mouse: $25
-  - USB Cable: $12
-`,
-        isUpdate: false,
-        format: "markdown",
-      },
-    ],
-  },
-  {
-    name: "Code Block Operations",
-    examples: [
-      {
-        name: "Extract js code",
-        code: `select(.code.lang == "js")`,
-        markdown: `# Sample codes
-\`\`\`js
-console.log("Hello, World!");
-\`\`\`
-
-\`\`\`python
-print("Hello, World!")
-\`\`\`
-
-\`\`\`js
-console.log("Hello, World!");
-\`\`\`
-`,
-        isUpdate: false,
-        format: "markdown",
-      },
-      {
-        name: "Exclude code",
-        code: `select(!.code)`,
-        markdown: `# Sample codes
-\`\`\`js
-console.log("Hello, World!");
-\`\`\`
-
-Some text here.
-
-\`\`\`python
-print("Hello, World!")
-\`\`\`
-
-More text here.
-`,
-        isUpdate: false,
-        format: "markdown",
-      },
-      {
-        name: "Extract language name",
-        code: `.code.lang`,
-        markdown: `# Sample codes
-\`\`\`js
-console.log("Hello, World!");
-\`\`\`
-
-\`\`\`python
-print("Hello, World!")
-\`\`\`
-
-\`\`\`rust
-println!("Hello, World!");
-\`\`\`
-`,
-        isUpdate: false,
-        format: "markdown",
-      },
-      {
-        name: "Extract all languages",
-        code: `nodes | pluck(.code.lang)`,
-        markdown: `# Sample codes
-\`\`\`js
-console.log("Hello, World!");
-\`\`\`
-
-\`\`\`python
-print("Hello, World!")
-\`\`\`
-
-\`\`\`rust
-println!("Hello, World!");
-\`\`\`
-`,
-        isUpdate: false,
-        format: "markdown",
-      },
-    ],
-  },
-  {
-    name: "Link and MDX Operations",
-    examples: [
-      {
-        name: "Extract MDX",
-        code: `select(is_mdx())`,
-        markdown: `import {Chart} from './snowfall.js'
-import { isDarkMode } from '../../../textusm/frontend/src/ts/utils';
-export const year = 2023
-
-# Last year's snowfall
-
-In {year}, the snowfall was above average.
-
-<Chart color="#fcb32c" year={year} />
-`,
-        isUpdate: false,
-        format: "mdx",
-      },
-      {
-        name: "Extract link URL",
-        code: `.link.url`,
-        markdown: `# Links
-
-Here is a [link to GitHub](https://github.com).
-Another [link to documentation](https://docs.example.com).
-And a [relative link](./readme.md).
-`,
-        isUpdate: false,
-        format: "markdown",
-      },
-    ],
-  },
-  {
-    name: "Advanced Markdown Processing",
-    examples: [
-      {
-        name: "Markdown TOC",
-        code: `.h
-| let link = to_link("#" + to_text(self), to_text(self), "")
-| let level = .h.depth
-| if (not(is_none(level))): to_md_list(link, level)`,
-        markdown: `# [header1](https://example.com)
-
-- item 1
-- item 2
-
-## header2
-
-- item 1
-- item 2
-
-### header3
-
-- item 1
-- item 2
-
-#### header4
-
-- item 1
-- item 2`,
-        isUpdate: false,
-        format: "markdown",
-      },
-      {
-        name: "Generate sitemap",
-        code: `def sitemap(item, base_url):
-  let path = replace(to_text(item), ".md", ".html")
-  | let loc = base_url + path
-  | s"<url>
-  <loc>\${loc}</loc>
-  <priority>1.0</priority>
-</url>"
-end
-| .[]
-| sitemap("https://example.com/")`,
-        markdown: `# Summary
-
-- [Chapter1](chapter1.md)
-- [Chapter2](Chapter2.md)
-  - [Chapter3](Chapter3.md)
-- [Chapter4](Chapter4.md)
-`,
-        isUpdate: false,
-        format: "markdown",
-      },
-    ],
-  },
-  {
-    name: "Custom Functions and Programming",
-    examples: [
-      {
-        name: "Custom function",
-        code: `def snake_to_camel(x):
-  let words = split(x, "_")
-  | foreach (word, words):
-      let first_char = upcase(first(word))
-      | let rest_str = downcase(word[1:len(word)])
-      | s"\${first_char}\${rest_str}";
-  | join("")
-end
-| snake_to_camel()`,
-        markdown: `# sample_codes`,
-        isUpdate: false,
-        format: "markdown",
-      },
-      {
-        name: "Map function",
-        code: `map([1, 2, 3, 4, 5], fn(x): x * 2;)`,
-        markdown: `# numbers`,
-        isUpdate: false,
-        format: "markdown",
-      },
-      {
-        name: "Filter function",
-        code: `filter([1, 2, 3, 4, 5], fn(x): x > 3;)`,
-        markdown: `# numbers`,
-        isUpdate: false,
-        format: "markdown",
-      },
-    ],
-  },
-  {
-    name: "File Processing",
-    examples: [
-      {
-        name: "CSV to markdown table",
-        code: `include "csv" | csv_parse(true) | csv_to_markdown_table()`,
-        markdown: `Product, Category, Price, Stock
-Laptop, Electronics, $1200, 45
-Monitor, Electronics, $350, 28
-Chair, Furniture, $150, 73
-Desk,  Furniture, $200, 14
-Keyboard, Accessories, $80, 35
-`,
-        isUpdate: false,
-        format: "raw",
-      },
-      {
-        name: "JSON to markdown table",
-        code: `include "json" | json_parse() | json_to_markdown_table()`,
-        markdown: `
-    [
-      { "Product": "Laptop", "Category": "Electronics", "Price": "$1200", "Stock": 45 },
-      { "Product": "Monitor", "Category": "Electronics", "Price": "$350", "Stock": 28 },
-      { "Product": "Chair", "Category": "Furniture", "Price": "$150", "Stock": 73 },
-      { "Product": "Desk", "Category": "Furniture", "Price": "$200", "Stock": 14 },
-      { "Product": "Keyboard", "Category": "Accessories", "Price": "$80", "Stock": 35 }
-    ]
-`,
-        isUpdate: false,
-        format: "raw",
-      },
-    ],
-  },
-];
-
-// Flatten examples for backward compatibility
-const EXAMPLES = EXAMPLE_CATEGORIES.flatMap((category) => category.examples);
-
 export const Playground = () => {
   const [code, setCode] = useState<string | undefined>(
-    localStorage.getItem(CODE_KEY) ?? EXAMPLES[0].code
+    localStorage.getItem(CODE_KEY) ?? EXAMPLES[0].code,
   );
   const [markdown, setMarkdown] = useState<string | undefined>(
-    localStorage.getItem(MARKDOWN_KEY) ?? EXAMPLES[0].markdown
+    localStorage.getItem(MARKDOWN_KEY) ?? EXAMPLES[0].markdown,
   );
   const [isUpdate, setIsUpdate] = useState(
-    localStorage.getItem(IS_UPDATE_KEY) === "true"
+    localStorage.getItem(IS_UPDATE_KEY) === "true",
   );
   const [isEmbed, setIsEmbed] = useState(false);
   const [result, setResult] = useState("");
@@ -388,17 +57,17 @@ export const Playground = () => {
       return format === "markdown"
         ? "markdown"
         : format === "text"
-        ? "text"
-        : format === "mdx"
-        ? "mdx"
-        : format === "html"
-        ? "html"
-        : format === "raw"
-        ? "raw"
-        : format === "null"
-        ? "null"
-        : null;
-    })()
+          ? "text"
+          : format === "mdx"
+            ? "mdx"
+            : format === "html"
+              ? "html"
+              : format === "raw"
+                ? "raw"
+                : format === "null"
+                  ? "null"
+                  : null;
+    })(),
   );
   const [activeTab, setActiveTab] = useState<"output" | "ast">("output");
   const [astResult, setAstResult] = useState("");
@@ -425,7 +94,7 @@ export const Playground = () => {
     return storedValue === null ? true : storedValue !== "false";
   });
   const [saveStatus, setSaveStatus] = useState<"saved" | "saving" | "unsaved">(
-    "saved"
+    "saved",
   );
   const [deleteConfirmDialog, setDeleteConfirmDialog] = useState<{
     path: string;
@@ -487,7 +156,7 @@ export const Playground = () => {
 
       if (!supported) {
         console.warn(
-          "Origin Private File System is not supported in this browser"
+          "Origin Private File System is not supported in this browser",
         );
         return;
       }
@@ -508,7 +177,7 @@ export const Playground = () => {
           initialTabs.length > 0
         ) {
           const activeTab = initialTabs.find(
-            (tab: Tab) => tab.id === savedActiveTabId
+            (tab: Tab) => tab.id === savedActiveTabId,
           );
           if (activeTab) {
             try {
@@ -520,7 +189,7 @@ export const Playground = () => {
               console.error("Failed to restore active tab:", error);
               // If active tab file doesn't exist, clear it
               setTabs((prev) =>
-                prev.filter((tab) => tab.id !== savedActiveTabId)
+                prev.filter((tab) => tab.id !== savedActiveTabId),
               );
               setActiveTabId(null);
             }
@@ -616,7 +285,7 @@ export const Playground = () => {
           listStyle,
           linkTitleStyle,
           linkUrlStyle,
-        })
+        }),
       );
     } catch (e) {
       setResult((e as Error).toString());
@@ -674,7 +343,7 @@ export const Playground = () => {
       },
     };
     const compressed = LZString.compressToEncodedURIComponent(
-      JSON.stringify(shareData)
+      JSON.stringify(shareData),
     );
     const url = `${window.location.origin}${window.location.pathname}#${compressed}`;
     window.location.hash = compressed;
@@ -721,7 +390,7 @@ export const Playground = () => {
       const value = e.target.value;
       setListStyle(value as mq.Options["listStyle"]);
     },
-    []
+    [],
   );
 
   const handleChangeLinkUrlStyle = useCallback(
@@ -729,7 +398,7 @@ export const Playground = () => {
       const value = e.target.value;
       setLinkUrlStyle(value as mq.Options["linkUrlStyle"]);
     },
-    []
+    [],
   );
 
   const getFileType = (filePath: string): Tab["fileType"] => {
@@ -790,7 +459,7 @@ export const Playground = () => {
       setEditorContent(tab.filePath, tab.content);
       setSelectedFile(tab.filePath);
     },
-    [tabs]
+    [tabs],
   );
 
   const handleTabClose = useCallback(
@@ -801,7 +470,7 @@ export const Playground = () => {
       // Check if tab has unsaved changes
       if (tab.isDirty) {
         const shouldClose = window.confirm(
-          `"${tab.filePath}" has unsaved changes. Do you want to close it?`
+          `"${tab.filePath}" has unsaved changes. Do you want to close it?`,
         );
         if (!shouldClose) return;
       }
@@ -823,7 +492,7 @@ export const Playground = () => {
         }
       }
     },
-    [tabs, activeTabId]
+    [tabs, activeTabId],
   );
 
   const handleFileSelect = useCallback(
@@ -839,7 +508,7 @@ export const Playground = () => {
         alert(`Failed to read file: ${error}`);
       }
     },
-    [openOrSwitchToTab]
+    [openOrSwitchToTab],
   );
 
   const handleCreateFile = useCallback(
@@ -865,11 +534,11 @@ export const Playground = () => {
         alert(
           `Failed to create file: ${
             error instanceof Error ? error.message : String(error)
-          }`
+          }`,
         );
       }
     },
-    [loadFiles, openOrSwitchToTab]
+    [loadFiles, openOrSwitchToTab],
   );
 
   const handleCreateFolder = useCallback(
@@ -891,11 +560,11 @@ export const Playground = () => {
         alert(
           `Failed to create folder: ${
             error instanceof Error ? error.message : String(error)
-          }`
+          }`,
         );
       }
     },
-    [loadFiles]
+    [loadFiles],
   );
 
   const handleDeleteFile = useCallback((path: string) => {
@@ -971,7 +640,7 @@ export const Playground = () => {
         // Prevent converting directory to file or vice versa
         if (isDirectory && hasExtension && !currentHasExtension) {
           alert(
-            "Cannot rename a folder to a file name. Folders cannot have file extensions."
+            "Cannot rename a folder to a file name. Folders cannot have file extensions.",
           );
           setIsRenaming(false);
           return;
@@ -1006,8 +675,8 @@ export const Playground = () => {
         // Update tab path if the file is open in a tab
         setTabs((prev) =>
           prev.map((tab) =>
-            tab.filePath === path ? { ...tab, filePath: newPath } : tab
-          )
+            tab.filePath === path ? { ...tab, filePath: newPath } : tab,
+          ),
         );
 
         // If this was the current file, reopen it with the new path
@@ -1026,7 +695,7 @@ export const Playground = () => {
         alert(
           `Failed to rename: ${
             error instanceof Error ? error.message : String(error)
-          }`
+          }`,
         );
         // Reload files to ensure UI is in sync
         await loadFiles();
@@ -1034,7 +703,7 @@ export const Playground = () => {
         setIsRenaming(false);
       }
     },
-    [loadFiles, selectedFile, currentFilePath, code]
+    [loadFiles, selectedFile, currentFilePath, code],
   );
 
   const handleMoveFile = useCallback(
@@ -1054,7 +723,7 @@ export const Playground = () => {
         if (exists) {
           const targetLocation = targetPath ? targetPath : "root";
           alert(
-            `A file or folder named "${fileName}" already exists in "${targetLocation}"`
+            `A file or folder named "${fileName}" already exists in "${targetLocation}"`,
           );
           return;
         }
@@ -1086,8 +755,8 @@ export const Playground = () => {
         // Update tab path if the file is open in a tab
         setTabs((prev) =>
           prev.map((tab) =>
-            tab.filePath === sourcePath ? { ...tab, filePath: newPath } : tab
-          )
+            tab.filePath === sourcePath ? { ...tab, filePath: newPath } : tab,
+          ),
         );
 
         // If this was the current file, reopen it with the new path
@@ -1106,13 +775,13 @@ export const Playground = () => {
         alert(
           `Failed to move: ${
             error instanceof Error ? error.message : String(error)
-          }`
+          }`,
         );
         // Reload files to ensure UI is in sync
         await loadFiles();
       }
     },
-    [loadFiles, selectedFile, currentFilePath, code]
+    [loadFiles, selectedFile, currentFilePath, code],
   );
 
   const saveCurrentFile = useCallback(async () => {
@@ -1128,8 +797,8 @@ export const Playground = () => {
         prev.map((tab) =>
           tab.id === activeTabId
             ? { ...tab, content: code, savedContent: code, isDirty: false }
-            : tab
-        )
+            : tab,
+        ),
       );
     } catch (error) {
       console.error("Failed to save file:", error);
@@ -1152,7 +821,7 @@ export const Playground = () => {
             return { ...tab, content: code, isDirty };
           }
           return tab;
-        })
+        }),
       );
     }
   }, [code, currentFilePath, isRenaming, activeTabId]);
@@ -1161,7 +830,7 @@ export const Playground = () => {
   useEffect(() => {
     localStorage.setItem(
       "mq-playground.sidebar-visible",
-      String(isSidebarVisible)
+      String(isSidebarVisible),
     );
   }, [isSidebarVisible]);
 
@@ -1221,7 +890,7 @@ export const Playground = () => {
               endColumn: error.endColumn,
               message: error.message,
               severity: monaco.MarkerSeverity.Error,
-            }))
+            })),
           );
         }
       });
@@ -1231,7 +900,7 @@ export const Playground = () => {
       triggerCharacters: [" ", "|", ":"],
       provideCompletionItems: async (
         model: editor.ITextModel,
-        position: IPosition
+        position: IPosition,
       ) => {
         const wordRange = model.getWordUntilPosition(position);
 
@@ -1252,10 +921,10 @@ export const Playground = () => {
                 value.valueType === "Function"
                   ? monaco.languages.CompletionItemKind.Function
                   : value.valueType === "Variable"
-                  ? monaco.languages.CompletionItemKind.Variable
-                  : value.valueType === "Selector"
-                  ? monaco.languages.CompletionItemKind.Method
-                  : monaco.languages.CompletionItemKind.Property,
+                    ? monaco.languages.CompletionItemKind.Variable
+                    : value.valueType === "Selector"
+                      ? monaco.languages.CompletionItemKind.Method
+                      : monaco.languages.CompletionItemKind.Property,
               insertText:
                 value.valueType === "Function"
                   ? `${value.name}(${
@@ -1275,7 +944,7 @@ export const Playground = () => {
                 endColumn: wordRange.endColumn,
               },
             };
-          }
+          },
         );
 
         const snippets: languages.CompletionItem[] = [
@@ -1627,7 +1296,7 @@ export const Playground = () => {
                           const globalIndex =
                             EXAMPLE_CATEGORIES.slice(0, categoryIndex).reduce(
                               (acc, cat) => acc + cat.examples.length,
-                              0
+                              0,
                             ) + exampleIndex;
                           return (
                             <option key={globalIndex} value={globalIndex}>
